WIP: start new session concept
